Changed libraries + test dice command
I changed from discord.py to py-cord due to various support concerns arisen from the other users of the library.

To confirm it worked, i had written a dice rolling command that returns a value based on the die's that the user had selected.

commands.py was deleted for maintenances sake, so that each command could be in its own file.